fix compiler issue

+++ src/frames.rs
@@ -210,7 +210,7 @@ impl From<UnresolvedFrames> for Frames {
         while let Some(frame) = frame_iter.next() {
             let mut symbols: Vec<Symbol> = Vec::new();
 
-            if let Some(perfmap_symbol) = resolve_in_perfmap(frame.ip()) {
+            if let Some(perfmap_symbol) = resolve_in_perfmap(frame.ip() as usize) {
                 symbols.push(perfmap_symbol);
             } else {
                 frame.resolve_symbol(|symbol| {
